Actualizacion para uso de observables y gestion de mensajes co servicios

+++ src/app/listado-productos/formulario-producto/formulario-producto.ts
@@ -24,10 +24,11 @@ export class FormularioProducto {
   this.descripcionProducto = "";
   this.precioProducto = 0;
   }
+  
 
   agregarNuevoProducto(event: Event) {
     //event.preventDefault();
-    if (this.descripcionProducto != null && this.precioProducto != null && !isNaN(this.precioProducto)) {
+    if (this.descripcionProducto != null && this.descripcionProducto != "" && this.precioProducto != null && !isNaN(this.precioProducto) && (this.precioProducto > 0)){
       const producto = new ProductoModelo(this.descripcionProducto, this.precioProducto);
       this.servicioProductos.addNuevoProducto(producto);
       this.eliminarValoresProducto();
